test_shop/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
stripe==5.1.1
python-dotenv
gunicorn==20.1.0
psycopg2-binary